Added Very Basic "Category Dividers" Button, 10/18/2025
Added Very Basic "Category Dividers" Button. What this does is add instead of a new entry to the ingredient list, it adds a "spacer" entry basically that had a large font and a place to put a name for a sub-category in a recipe or whatever. Say if you wanted 2 different sections for the ingredients, 1 for "Cake Batter" the other for "Frosting" or something like that, these "Spacer" entries allow for doing that, might polish this a slight bit more next time, if not I'll just think of and start work on the next feature, will see, 10/18/2025.