Mouse Cursor Now Follows Reorder Buttons, 10/19/2025
Mouse Cursor Now Follows Reorder Buttons. All this does is make reordering items be quicker, since you don't have to reposition your mouse to the new position of the button you just clicked everytime you change the order, it "follows" it after being clicked. The only drawback currently is that if the button scrolled off the scrollContainer's visible space, it won't automatically scroll to compensate, tried it but it didn't work great, so just decided to not bother for now, 10/19/2025.